Handle disconnected popup in passwordless login.
The passwordless login method sends the user a link in an email. When
the user follows this link a new browser window opens.

However the original popup window will not redirect and remains open
with a page allowing to resend the link.

The connect-js library was changed to handle this case so that it would
expect the page which is opened when the email link is followed
to handle the callback so that the session is populated appropriately
in localStorage.

connect-js now listens for that and then closes the popup window.
It also expect the email page to close itself.

This may not work for all browsers.  I did see this work in Chrome and Safari
but fail in Firefox.
Note that the popup Callback should behave in the same way it it is
redirected by the server, which happens when other authentication
methods are used.

This also adds the newer dependencies and improves logging similar to
how it was done in connect-js.