pull app configuration into app.config
Files in /app.config/ are now generated on grunt config
based on the templates in /app.config.templ/* and the
corresponding variables in /authconfig.json
authconfig.json contains variables describing the
issuer (connect server) and the server the app will run
on as wells as other registration information.

After `grunt config` there are 2 files in /app.config:

1. /app.config/anvil-config.js which is used by app/scripts/*.js files
to configure the Anvil client library.

2. /app.config/register_with_anvil_connect.sh is a shell
script which registers this client with the server using the `nvl`
command line.

This change allows to to have to eliminate intermediate copies
of all app files which where previously in app.browser.

Now /app.browser is used primarily to save dev-bundles generated
by browserify. See npm run watchify to keep these up to date.
In addition index.html still uses config values for declaring the
session management iframes in particular for the op
iframe.